Fix reward being returned early after the first episode
With commit e4f133bc5413b3e5f9c1999167e5f5a351106bfb, `train` was now
returning the mean reward but for the A2C and DQN agents the reward is
being returned prematurly. Only one episode runs and with this change
the mean reward is returned after all the steps and episodes are
completed.